model extractor in different file + adding of seq2seq pred in extractor
